Create MazeGen Version 2
This is the second version of MazeGen. Here the backtracking was working and as far as I could tell the algorithm should have worked flawlessly. However once the maze went above a size of 73*73 a stack overflow would occur. This was purely because the recursion was so deep and so much was stored on the stack. This version is what helped me realise that recursion was not the answer to my maze generation problem.