Create MazeGen Version 3
This is the fully working version of maze gen. I rewrote the "Algorithm" class so that it utilises a while loop rather than recursion. This meant that the maze could be as large as I wanted. This version is however still a console application meaning the maze where limited by what I could display in the console.